HSP analysis by bitscore, skip human & predicted results
Added HSP analysis of the blast file to extract top 10 results based on
HSP bit score. The shell script also now asks if you want to include
genomic sequences from humas or predicted sequences and acts
accordingly. Gives printout of the names of each of the 10 sequences
that best match